requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
python-slugify==8.0.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
